+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # MobX-React Changelog
 
+### 5.3.0
+
+* Added `disposeOnUmount` utility / decorator to call disposable properties (reaction, autorun, etc) automatically on `componentWillUnmount`
+* Introduced new method to patch lifecycle methods which should be more compatible with for example arrow functions.
+
 ### 5.2.8
 
 * Make sure `mobx-react` doesn't require `Object.assign` polyfill